requirements.txt
streamlit
pandas
scikit-learn
joblib
Pillow
numpy